urfube/requirements.txt
aiojobs==1.1.0
anyio==3.6.2
async-timeout==4.0.2
bcrypt==4.0.1
boto3==1.26.109
botocore==1.29.109
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.1
ecdsa==0.18.0
fastapi==0.95.0
fastapi-jsonrpc==2.4.1
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
jmespath==1.0.1
packaging==23.0
passlib==1.7.4
peewee==3.16.0
pluggy==1.0.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.7
pytest==7.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
rfc3986==1.5.0
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
watchfiles==0.19.0
websockets==11.0.1
